feat: address

+++ projects/server/src/controllers/category.js
@@ -11,17 +11,23 @@ const categoryController = {
       const check = await db.Category.findOne({ where: { name } });
 
       if (check) {
-        fs.unlinkSync(req.file.path);
+        if (filename) {
+          fs.unlinkSync(`${__dirname}/../public/category/${filename}`);
+        }
         return res.status(400).send({ message: "name alrdy exist" });
       }
+
       await db.Category.create(
-        { name, category_img: CATEGORY_URL + filename },
+        { name, category_img: filename ? "category/" + filename : null },
         { transaction: t }
       );
+
       await t.commit();
       return res.status(200).send({ message: "success add Category" });
     } catch (err) {
-      fs.unlinkSync(req.file.path);
+      if (filename) {
+        fs.unlinkSync(`${__dirname}/../public/category/${filename}`);
+      }
       await t.rollback();
       return res.status(500).send(err.message);
     }
@@ -80,33 +86,44 @@ const categoryController = {
   },
   editCategory: async (req, res) => {
     const t = await db.sequelize.transaction();
+    const filename = req?.file?.filename;
     try {
-      const { name, category } = req.body;
-      const filename = req?.file?.filename || category;
+      const { name } = req.body;
       const check = await db.Category.findOne({ where: { id: req.params.id } });
 
-      if (check?.dataValues?.category_img) {
-        fs.unlinkSync(
-          `${__dirname}/../public/category/${
-            check.dataValues.category_img.split("/")[5]
-          }`
-        );
+      if (check) {
+        if (filename) {
+          fs.unlinkSync(`${__dirname}/../public/category/${filename}`);
+        }
+        return res.status(400).send({ message: "name alrdy exist" });
       }
 
       await db.Category.update(
         {
           name,
-          category_img: !req?.file?.filename
-            ? category
-            : CATEGORY_URL + filename,
+          category_img: filename
+            ? "category/" + filename
+            : check?.dataValues?.category_img || null,
         },
-        { where: { id: req.params.id } },
-        { transaction: t }
+        { where: { id: req.params.id }, transaction: t }
       );
+
+      if (check?.dataValues?.category_img) {
+        if (filename) {
+          fs.unlinkSync(
+            `${__dirname}/../public/category/${
+              check.dataValues.category_img.split("/")[1]
+            }`
+          );
+        }
+      }
+
       await t.commit();
       return res.status(200).send({ message: "success edit category" });
     } catch (err) {
-      fs.unlinkSync(req.file.path);
+      if (filename) {
+        fs.unlinkSync(`${__dirname}/../public/category/${filename}`);
+      }
       await t.rollback();
       return res.status(500).send(err.message);
     }
@@ -117,8 +134,7 @@ const categoryController = {
       const { name } = req.body;
       await db.SubCategory.update(
         { name },
-        { where: { id: req.params.id } },
-        { transaction: t }
+        { where: { id: req.params.id }, transaction: t }
       );
       await t.commit();
       return res.status(200).send({ message: "success edit subcategory" });
@@ -132,23 +148,23 @@ const categoryController = {
     try {
       const check = await db.Category.findOne({ where: { id: req.params.id } });
 
-      await db.SubCategory.destroy(
-        { where: { category_id: req.params.id } },
-        { transaction: t }
-      );
+      await db.SubCategory.destroy({
+        where: { category_id: req.params.id },
+        transaction: t,
+      });
+
+      await db.Category.destroy({
+        where: { id: req.params.id },
+        transaction: t,
+      });
 
       if (check?.dataValues?.category_img) {
         fs.unlinkSync(
           `${__dirname}/../public/category/${
-            check.dataValues.category_img.split("/")[5]
+            check.dataValues.category_img.split("/")[1]
           }`
         );
       }
-
-      await db.Category.destroy(
-        { where: { id: req.params.id } },
-        { transaction: t }
-      );
 
       await t.commit();
       return res.status(200).send({ message: "success delete category" });
@@ -160,10 +176,10 @@ const categoryController = {
   deleteSubcategory: async (req, res) => {
     const t = await db.sequelize.transaction();
     try {
-      await db.SubCategory.destroy(
-        { where: { id: req.params.id } },
-        { transaction: t }
-      );
+      await db.SubCategory.destroy({
+        where: { id: req.params.id },
+        transaction: t,
+      });
       await t.commit();
       return res.status(200).send({ message: "success delete subcategory" });
     } catch (err) {
